♻️ refactor: 아이템 선택하는 상태 로직 분리
