+++ src/app/likes/components/index.ts
@@ -1,0 +1,1 @@
+export { default as LikesList } from './LikesList';
